+++ tests/auxiliary/json.rs
@@ -129,7 +129,7 @@ mod tests {
         .unwrap()
         .filter_map(Result::ok)
         .collect();
-        assert_eq!(files.len(), 11);
+        assert_eq!(files.len(), 13);
 
         for file in files {
             let s = fs::read_to_string(file).unwrap();